Split package 'ui.pbem' into 'ui.posted.game.pbf' and 'ui.postged.game.pbem'
