+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/posted/game/pbf/PbemSetupPanel.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.startup.ui.pbem;
+package games.strategy.engine.framework.startup.ui.posted.game.pbf;
 
 import com.google.common.base.Preconditions;
 import games.strategy.engine.data.GameData;
@@ -11,6 +11,10 @@ import games.strategy.engine.framework.startup.mc.HeadedLaunchAction;
 import games.strategy.engine.framework.startup.ui.PlayerSelectorRow;
 import games.strategy.engine.framework.startup.ui.PlayerType;
 import games.strategy.engine.framework.startup.ui.SetupPanel;
+import games.strategy.engine.framework.startup.ui.posted.game.DiceServerEditor;
+import games.strategy.engine.framework.startup.ui.posted.game.pbem.EmailSenderEditor;
+import games.strategy.engine.framework.startup.ui.posted.game.pbem.ForumPosterEditor;
+import games.strategy.engine.framework.startup.ui.posted.game.pbem.ForumPosterEditorViewModel;
 import games.strategy.engine.posted.game.pbf.IForumPoster;
 import games.strategy.engine.random.PbemDiceRoller;
 import java.awt.Dialog;
