+++ game-core/src/main/java/games/strategy/triplea/settings/SelectionComponentFactory.java
@@ -4,7 +4,7 @@ import static org.triplea.util.Arrays.withSensitiveArray;
 import static org.triplea.util.Arrays.withSensitiveArrayAndReturn;
 
 import com.google.common.base.Strings;
-import games.strategy.engine.framework.startup.ui.pbem.DiceServerEditor;
+import games.strategy.engine.framework.startup.ui.posted.game.DiceServerEditor;
 import games.strategy.engine.framework.system.HttpProxy;
 import games.strategy.engine.posted.game.pbem.IEmailSender;
 import java.awt.Component;
